requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1